refactor: 데이터 로더 구조 개선 및 추상 클래스 도입
