+++ data_loaders/__init__.py
@@ -1,1 +1,2 @@
-from .data_loader import *
+from .inference_data_loader import *
+from .train_data_loader import *
